Homework3
