+++ Homework2/Homework3cs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Homework
+{
+    public class Homework3cs
+    {
+        public int id = 123; 
+        public string name = "Alexandra";
+    }
+}
