+++ Homework2/bin/Debug/net6.0/Homework.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v6.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v6.0": {
+      "Homework/1.0.0": {
+        "runtime": {
+          "Homework.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "Homework/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
