+++ Homework2/bin/Debug/net6.0/Homework.exe
