+++ Homework2/bin/Debug/net6.0/Homework.runtimeconfig.json
@@ -1,0 +1,9 @@
+{
+  "runtimeOptions": {
+    "tfm": "net6.0",
+    "framework": {
+      "name": "Microsoft.NETCore.App",
+      "version": "6.0.0"
+    }
+  }
+}
